Optimization: TryEstimateFeeRate should not catch exceptions
One of the advantages of using a TryFoo method is that
the developer can not only suppose that it doesn't throw
exceptions, but also that it doesn't generate them under
the hood, so that this alternative is faster.

Therefore this commit inverts the recently committed logic:
the implementation that returns null is the one privately
shared by all the other public-facing methods, and only the
non-Try ones have logic to throw exceptions.