Transaction: make signing idiot-proof
If transaction.Sign(key) was used, one could have
assumed that he was using an overload in which the
second parameter `bool assumeP2SH` would be optional.
However, there was no overload with a default value
for the second bool parameter; but there is an overload
that takes a `params` keyword for the second parameter,
which means that it can accept an empty array. But being
empty meant that it was not signing at all, even if
the method Sign() was called.

Make it idiot-proof so that if an empty array is passed,
it throws an ArgumentException. On top of that, make
Check() function also check if the transaction is signed.

Fixes #257