Transaction: make AddInput(Transaction,int) be more clever
I assume this method was added to make it easier to add
inputs to a transaction without the need to create the
Input object itself. This is great, but to make it even
more great, we can even fill the ScriptSig property of
the new input for the API user so that they don't need
to mutate the inputs of the transaction after returning
it in order to be able to sign it later.

(If the ScriptSig of the input returned by this method
was not modified before trying to sign the transaction,
signing it would throw the exception below:)

```
System.InvalidOperationException: ScriptSigs should be filled with either previous
scriptPubKeys or redeem script (for P2SH)
```