Transaction: make Sign(key,coin) overloads stop using `params` keyword
Having the params keyword meant that it was easy to shoot
yourself in the foot, as a newbie, by not providing any
element of the array at all, which would end up not signing
the transaction at all without throwing any exception (silent
failing). See discussion in:
https://github.com/MetacoSA/NBitcoin/pull/260

Also provide two new overloads for non-array variables, though
(this is harmless, as opposed to params keyword, because
we make sure at least there's one element provided).

As this is an API breaking change, version is bumped to 5.0