TX Inputs to be RBF-enabled by default
When creating an input, there was a need to explicitly
use the Sequence property setter in order to enable RBF
on it.

Bitcoin Core has recently enabled RBF by default even
from the GUI [1], so it should be safe to assume that
this feature is stable enough to be enabled by default
in NBitcoin too.

The implementation to make this possible needs to use
a Nullable type for the backing field because the C#
compiler doesn't allow structs with parameterless ctors
or default values to be assigned to their fields.

[1] https://bitcoin.org/en/release/v0.16.0