RestClient keeps a network reference (#384)
It is possible to specify the Network instance that can be used by
a RestClient instance. That network is passed as an argument in the
`contructor` (even when it is not required for achieve the current
responsability of the class).

Backward compatibility is achieve by keeping the existing constructor
and setting the Main network by default.