Creates interface for abstracting VerifyMessage
A new `IPubkeyHashUsable` interface is introduced in order to
have a common abstraction for `BitcoinWitPubkeyAddress` and
`BitcoinPubkeyAddress` address type which can verify messages.

The name of the interface is probably not good enough. It is
also possible that you find a better way to implement the same.
The idea of this commit is start a discussion to agree the best
way to implement it, if appropiated.

The next commit (an unit test) shows how it could be used.