New GetRawTransactionInfo method
This PR adds a new `GetRawTransactionInfo` method to the `RpcClient`
class because the exisiting `GetRawTransaction` method returns simply a
`Transaction` instance and sometimes developers also need useful
informations such as `BlochHash` and/or `Confirmations` for checking
whether the tx is in the blockchain or not; in which block it is and how
mature it is.

Note this new method returns redundant information that can
be extracted directly from the also-returned transaction. IMO this is
not something wrong (I can remove it too)

Having two methods with similar names and porpouse doesn't sound a good
idea IMO, maybe we could mark `GetRawTransaction(Async)` as `Obsolete`
and remove them in some version in the future.