Update bip 158 test vectors
The bip 158 test vectors are now in json format (using the original
file) and there is a new tested scenario. Additionally, the vectors
use the new serialization which consists of N as a varint. This means
an empty filter is serialized as "00" instead of "" (empty).

By last, there were a missing changes after the dotnet update to 2.1
that are fixed here too.