Crypto.Pbkdf2: make class public (but obsolete)
In the same way SCrypt is already public, having
Pbkdf2 be public can also be useful (for example
to implement the WarpWallet algorithm without
depending on CryptSharp).

Note: changing some methods from public to
internal is necessary because IMac interface
is internal (this is not a breaking change,
as they were being public before, but inside
an internal class, therefore not really public).

The reason of why this is marked as Obsolete
from the beginning is that NBitcoin might adopt
Rfc2898DeriveBytes class instead (if/when it
can raise the requirement of .NET 4.7.2, because
this is the .NET profile version where the
overloads that accept the hashing algorithm
are introduced).