Fix ifdefs: .NET Core != .NET Standard
As discussed in https://github.com/MetacoSA/NBitcoin/pull/444#issuecomment-425695831
we shouldn't be mixing up these ifdefs.

NOTE: the CryptSharp stuff needs to be #if-wrapped with NETCORE, even if
those APIs are present in .NetCore v2 already, because otherwise the test
BIP38Tests..EncryptedSecretECmultiplyNoLotSimple generates a
StackOverflowException that takes the whole suite down.