Nits and bump (#640)
* XML docs nitpicks

* Money: use correct order of arguments when using ArgumentOutOfRangeEx

It's a bit crazy but for ArgumentOutOfRangeException ctor,
the paramName is the first parameter[1], even if it is the
other way around for other similar exceptions (e.g. [2]).

[1] https://docs.microsoft.com/en-us/dotnet/api/system.argumentoutofrangeexception.-ctor?view=netframework-4.7.2#System_ArgumentOutOfRangeException__ctor_System_String_System_String_
[2] https://docs.microsoft.com/en-us/dotnet/api/system.argumentexception.-ctor?view=netframework-4.7.2#System_ArgumentException__ctor_System_String_System_String_

* Use nameof keyword in more places

* Money: add readonly keyword and avoid int-related warning

* Money: remove unnecessary cast

* Public setters are evil

No need for them in this case (there's already a ctor that sets them)

* bump

geewallet needs the new SendAll API introduced recently,
so new nuget package would be appreciated.