TransactionBuilder: stop having RBF disabled by default
Old CreateTransactionBuilder() methods are now marked as Obsolete; the
new ones, introduced in this commit, will force the NBitcoin consumer
to make a decision about what kind of transactions will be created by
her/his TransactionBuilder instance:

* if using EnhanceOnChainPrivacy mode, RBF will be disabled and the
inputs & outputs will be shuffled;
* if using Lightning mode, RBF will be enabled (ReplaceByFee is very
important in this environment because fee rates might change a lot
since the moment when a channel was opened until when it needs to
be closed), and shuffling inputs or outputs will be disabled.

Why disabling shuffling for Lightning? Because according to the spec,
they need to be ordered lexicographically (e.g. the previous defaults
were causing a bug in DotNetLightning: [1]).

[1] https://github.com/joemphilips/DotNetLightning/pull/133