Signal sendaddrv2 and process signal message
Immediatelly after receiving the remote node's `VerAck` we send
a `sendaddrv2` message to let the peer to know the client knows
handle `addrv2` messages.

When the node receives the `sendaddrv2` message it sets the
`PreferAddressV2` to true. That flag is used to decide what
type of messages to send (`addr` or `addrv2`)

In case the remote node doesn't signal support for `addrv2`
and we have addresses that are not compatible with v1, we
filter all those network addresses and send only those that
were supported by v1.