New addrv2 and sendaddrv2 signal messages
Add two new messages:

* `addrv2` that can contain new address types (torv3, cjdns and i2p).
* `sendaddrv2` to signal the preference for `addrv2` messages, what means we
understand and can process the new address types).

Note 1: `addrv2` are **not** serialized exactly in the same way that `addr`
Note 2: in this PR we are going to send `sendaddrv2` as specified in the
bip155 immediatly after `VerAck` but that will change soon and we will
need to send it before `VerAck` otherwise the node will disconnect us.