add ecash
use MEF to export all Altcoins and
remove the requirement to add Altcoins to
NBitcoin in the future.

Instead Altcoins will need to export `INetworkSet`.
It will be up to consumers (like NBXplorer and BTCPayServer)
to discover other Altcoins using either an Altcoin assembly directory
or a plugin "installation" mechanism.