Refactor Bitcoin networks and support MutinyNet
This PR refactors Network and its Bitcoin networks, in that it pulls them to the NetworkSet class to reduce its coupling and make their instantiation similar to how every other network is created. It is IMO a much cleaner result.

Additioanlly, it adds support for MutinyNet, a custom signet network with 30s blocks and many covenant soft forks activated. Most of LN testing infra runs on Mutinynet, and is very useful.